==> 1. Glaucoma - Mat to CSV.ipynb <==
from scipy.io import loadmat
import pandas as pd
import numpy as np

filename = 'C:\Luna_CS\Aravind\ORIGA\OrigaList.mat'
mat = loadmat(filename)

origa = mat['Origa']

data = {}
data['Eye'] = []
data['Filename'] = []
data['ExpCDR'] = []
data['Set'] = []
data['Glaucoma'] = []
for i in range(650):
    row = origa[:, i][0]
    data['Eye'].append(row[0][0])
    data['Filename'].append(row[1][0])
    data['ExpCDR'].append(row[2][0][0])
    data['Set'].append(row[3][0])
    data['Glaucoma'].append(row[4][0][0])
data = pd.DataFrame(data)
data.head()

data.to_csv('C:\Luna_CS\Aravind\ORIGA\OrigaList.csv', index=False)
