==> 2. Baseline Model - No Cropping.ipynb <==
import tensorflow as tf
from tensorflow import keras
from tensorflow.keras.preprocessing import image
from tensorflow.keras.applications.resnet50 import ResNet50
import tensorflow.keras.layers as L
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

IMAGE_DIR = '../input/origalist/Images/Images'

df = pd.read_csv('../input/origalist/OrigaList.csv')
df.drop(['Eye', 'ExpCDR', 'Set'], axis=1, inplace=True)
df.head()

data_gen = image.ImageDataGenerator(rescale=1./255,
                                   zoom_range=0.3,
                                   rotation_range=20,
                                   validation_split=0.3,
                                   width_shift_range=0.3)

train_generator = data_gen.flow_from_dataframe(df,
                                              IMAGE_DIR,
                                              x_col='Filename',
                                              y_col='Glaucoma',
                                              target_size=(256, 256),
                                              class_mode='raw',
                                              batch_size=16,
                                              shuffle=True,
                                              subset='training')

val_generator = data_gen.flow_from_dataframe(df,
                                            IMAGE_DIR,
                                            x_col='Filename',
                                            y_col='Glaucoma',
                                            target_size=(256, 256),
                                            class_mode='raw',
                                            batch_size=8,
                                            subset='validation')

sum(df['Glaucoma'])/len(df['Glaucoma'])

model = keras.Sequential()
model.add(ResNet50(input_shape=(256, 256, 3), include_top=False, weights='imagenet'))
model.add(L.GlobalAveragePooling2D())
model.add(L.Dropout(0.4))
model.add(L.Dense(1, activation='sigmoid'))

model.compile(optimizer='adam', loss='binary_crossentropy', metrics=['acc', keras.metrics.AUC()])

history = model.fit(train_generator, steps_per_epoch=train_generator.n//train_generator.batch_size,
                   validation_data=val_generator, validation_steps=val_generator.n//val_generator.batch_size,
                   epochs=8)
